requirements.txt
aiohttp-jinja2==1.2.0
aiohttp==3.6.2
beautifulsoup4==4.8.1
jinja2==2.10.3
networkx
nltk==3.4.5
pyaml
pytest-aiohttp
stix2==1.2.1
stix2-patterns==1.2.1
taxii2-client==0.5.0
Sphinx==2.4.4
sphinx_rtd_theme==0.4.3
recommonmark
pyaml
scikit-learn==0.23.1
node2vec>=0.3.2
tqdm
spacy==2.2.4
feedparser
aiohttp_security
aiohttp_session
cryptography
pdfplumber==0.5.22
forestci
htmldate
pandas